backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
openai>=1.0.0
